+++ Forge/src/main/java/com/christofmeg/brutalharvest/client/data/BrutalBlockStateProvider.java
@@ -29,15 +29,15 @@ public class BrutalBlockStateProvider extends BlockStateProvider {
     @Override
     protected void registerStatesAndModels() {
 
-        makeCrop(BlockRegistry.TOMATO.get(), TomatoCropBlock.AGE);
+        makeCrop(BlockRegistry.TOMATO.get(), TomatoCropBlock.AGE, modLoc("block/lowered_cross"));
         makeCrop(BlockRegistry.LETTUCE.get(), LettuceCropBlock.AGE);
-        makeDoubleCrop(BlockRegistry.CORN.get(), 8, 2);
-        makeDoubleCrop(BlockRegistry.CUCUMBER.get(), 7, 4);
+        makeDoubleCrop(BlockRegistry.CORN.get(), 8, 2, modLoc("block/lowered_cross"));
+        makeDoubleCrop(BlockRegistry.CUCUMBER.get(), 7, 4, modLoc("block/lowered_cross"));
 
-        makeCrop(BlockRegistry.COTTON.get(), CottonCropBlock.AGE);
-        makeDoubleCrop(BlockRegistry.RAPESEED.get(), 8, 4);
+        makeCrop(BlockRegistry.COTTON.get(), CottonCropBlock.AGE, modLoc("block/lowered_cross"));
+        makeDoubleCrop(BlockRegistry.RAPESEED.get(), 8, 4, modLoc("block/lowered_cross"));
         makeCrop(BlockRegistry.SUGAR_BEET.get(), SugarBeetCropBlock.AGE);
-        makeCrop(BlockRegistry.STRAWBERRY.get(), StrawberryCropBlock.AGE);
+        makeCrop(BlockRegistry.STRAWBERRY.get(), StrawberryCropBlock.AGE, modLoc("block/lowered_cross"));
     //    makeCrop(BlockRegistry.ONION.get(), OnionCropBlock.AGE);
         makeBush(BlockRegistry.BLUEBERRY.get(), BlueberryBushBlock.AGE);
     }
@@ -52,6 +52,17 @@ public class BrutalBlockStateProvider extends BlockStateProvider {
         });
     }
 
+    public void makeCrop(Block block, IntegerProperty ageProperty, ResourceLocation customModel) {
+        String name = block.getDescriptionId().replace("block.brutalharvest.", "");
+        getVariantBuilder(block).forAllStates(state -> {
+            String modelName = name + "_stage" + state.getValue(ageProperty);
+            ResourceLocation textureLocation = new ResourceLocation(CommonConstants.MOD_ID, "block/" + modelName);
+            ConfiguredModel model = new ConfiguredModel(models().withExistingParent(modelName, customModel).texture("cross", textureLocation).renderType(CUTOUT));
+            return new ConfiguredModel[]{model};
+        });
+    }
+
+    @SuppressWarnings("unused")
     public void makeDoubleCrop(Block block, int growthStages, int firstStageWithLowerUpper) {
         String name = block.getDescriptionId().replace("block.brutalharvest.", "");
         if (block instanceof BaseDoubleCropBlock baseCropBlock) {
@@ -67,6 +78,26 @@ public class BrutalBlockStateProvider extends BlockStateProvider {
                 }
                 ResourceLocation textureLocation = new ResourceLocation(CommonConstants.MOD_ID, "block/" + modelName);
                 ConfiguredModel model = new ConfiguredModel(models().crop(modelName, textureLocation).renderType(CUTOUT));
+                return new ConfiguredModel[]{model};
+            });
+        }
+    }
+
+    public void makeDoubleCrop(Block block, int growthStages, int firstStageWithLowerUpper, ResourceLocation customModel) {
+        String name = block.getDescriptionId().replace("block.brutalharvest.", "");
+        if (block instanceof BaseDoubleCropBlock baseCropBlock) {
+            getVariantBuilder(block).forAllStates(state -> {
+                int age = baseCropBlock.getAge(state);
+                String modelName = name + "_stage" + age;
+                if (age >= firstStageWithLowerUpper) {
+                    if (age < growthStages) {
+                        modelName = name + "_stage" + age + "_lower";
+                    } else {
+                        modelName = name + "_stage" + (age - growthStages + firstStageWithLowerUpper) + "_upper";
+                    }
+                }
+                ResourceLocation textureLocation = new ResourceLocation(CommonConstants.MOD_ID, "block/" + modelName);
+                ConfiguredModel model = new ConfiguredModel(models().withExistingParent(modelName, customModel).texture("cross", textureLocation).renderType(CUTOUT));
                 return new ConfiguredModel[]{model};
             });
         }