+++ Forge/changelog.md
@@ -9,3 +9,4 @@
 * Fix Water Bottle recipe not showing with nbt in recipe viewers
 * Fix bug where no blocks could be placed against Blueberry Bushes
 * Made Blueberry bushes go back to age 0 instead of 2 when a knife is used on them
+* Fix block cross crop models on Corn, Cucumber, Cotton, Rapeseed, Strawberry and Tomato